refactor: address maintainer feedback on community section
