+++ src/locals/en.json
@@ -10,19 +10,16 @@
     "community": {
       "title_part1": "Built With ",
       "title_gradient": "Community",
-      "subtitle": "This project was born from a real problem and a desire to give back. Everyone is welcome — from first-time contributors to seasoned engineers.",
+      "subtitle": "This project was born from a real problem and a desire to give back. Everyone is welcome, from first time contributors to seasoned engineers.",
       "card_title": "Why I Built This",
-      "paragraphs": [
-        "I started this project after repeatedly running into the same frustration: old, unmaintained repositories with outdated READMEs and broken links. After opening countless issues to report them, I realized it would be better to solve the problem at scale.",
-        "So I built an open source tool focused on fixing broken links, with a strong emphasis on community and contributors. The goal is to grow step by step and help others along the way through code reviews and guidance.",
-        "The open source community has given me a lot — including a job referral — and this is my way of giving back. I'm always open to feedback and new contributors. If you'd like to contribute but aren't sure where to start, feel free to DM me. I'll personally help you find a good first issue. 🙌"
-      ],
+      "paragraph1": "I started this project after repeatedly running into the same frustration: old, unmaintained repositories with outdated READMEs and broken links. After opening countless issues to report them, I realized it would be better to solve the problem at scale.",
+      "paragraph2": "So I built an open source tool focused on fixing broken links, with a strong emphasis on community and contributors. The goal is to grow step by step and help others along the way through code reviews and guidance.",
+      "paragraph3": "The open source community has given me a lot (including a job referral) and this is my way of giving back. I'm always open to feedback and new contributors. If you'd like to contribute but aren't sure where to start, feel free to DM me. I'll personally help you find a good first issue. 🙌",
       "tech_stack_label": "Tech Stack",
       "tech_stack": {
         "typescript": "TypeScript",
         "react": "React",
-        "nodejs": "Node.js",
-        "base44": "Base44"
+        "nodejs": "Node.js"
       },
       "cta": "View on GitHub"
     }
