+++ src/pages/About/About.page.tsx
@@ -1,5 +1,5 @@
 import { AboutTitle } from './components/AboutTitle';
-import { CommunitySection } from './components/CommunitySection';
+import { CommunitySection } from './components/CommunitySection/CommunitySection';
 import { FeaturesCards } from './components/FeaturesCards';
 import { MissionCard } from './components/MissionCard';
 import { centerGrid, pageWrapperStyle } from './components/styles';
